reorganize mpi_parallel code into multiple modules
