+++ tenpy/mpi_parallel/__init__.py
@@ -1,5 +1,18 @@
 """Adjustments to employ MPI parallelization splitting up the MPO virtual legs.
 
+This module implements parallelization of DMRG with the MPI framework [MPI]_.
+It's based on the python interface of `mpi4py <https://mpi4py.readthedocs.io/>`_,
+which needs to be installed when you want to use classes in this module.
+
+.. note ::
+    This module is not imported by default, since just importing mpi4py already initializes MPI.
+    Hence, if you want to use it, you need to explicitly call
+    ``import tenpy.mpi_parallel`` in your python script.
+
+.. warning ::
+    MPI parallelization is still under active development.
+
+
 .. rubric:: Submodules
 
 .. autosummary::
@@ -8,6 +21,6 @@
 """
 # Copyright 2021 TeNPy Developers, GNU GPLv3
 
-from . import actions, simulation
+from . import helpers, actions, simulation
 
-__all__ = ["actions", "simulation"]
+__all__ = ["helpers", "actions", "simulation"]
